refactor(#3037): relocate thread_reuse PG/Discord helpers into services (backflow 24→22)
Move the Postgres/Discord-API thread-map helpers out of the route layer so
`src/services/` no longer reaches back into `crate::server`. The axum route
handlers stay in the server layer and consume the relocated helpers.

Relocated `server/routes/dispatches/thread_reuse.rs` →
`services/dispatches/discord_delivery/thread_reuse.rs`:
  - get_thread_for_channel_pg, get_mapped_thread_for_channel_pg,
    set_thread_for_channel_pg, set_thread_for_channel_map_only_pg,
    set_thread_for_channel_pg_with_active, clear_thread_for_channel_pg,
    lookup_thread_for_channel_from_map_pg, json_value_kind, try_reuse_thread,
    validate_channel_thread_maps_on_startup_with_backends (+ validation row/outcome).

Stay in server layer (axum handlers): get_card_thread, link_dispatch_thread,
get_pending_dispatch_for_thread.

Call-site updates:
  - services/.../discord_delivery/orchestration.rs: import the 5 thread-map
    helpers from the sibling `super::thread_reuse` instead of
    `crate::server::routes::dispatches::thread_reuse` (orchestration.rs 2→1).
  - services/discord/runtime_bootstrap.rs: re-point the startup thread-map
    validation call to the relocated services home (runtime_bootstrap.rs 1→0).

Behavior unchanged — pure relocation; every SQL statement, Discord-API probe,
and clear/reuse decision is byte-identical. The relocated file shrank below the
1000-line giant threshold, so its live `grandfathered` registry entry was
removed (frozen baseline path retained).

Gates: cargo fmt --check clean; cargo check --lib / --lib --tests clean (0 new
warnings); cargo test --lib discord_delivery (20 passed); backflow baseline
24→22 (0 regressions); generate_inventory_docs + check_agent_maintenance_docs
freshness passed; ci-script-checks.sh exit 0; multinode-transition.md Audited
touches updated.

Refs #3037
