+++ ARCHITECTURE.md
@@ -547,6 +547,7 @@ src/
 │   │   │   ├── guard.rs
 │   │   │   ├── mod.rs
 │   │   │   ├── orchestration.rs
+│   │   │   ├── thread_reuse.rs
 │   │   │   └── transport.rs
 │   │   ├── mod.rs
 │   │   ├── outbox_claiming.rs
