+++ scripts/audit_maintainability/baselines/service_server_backflow.json
@@ -3,7 +3,7 @@
   "rule": "service_server_backflow",
   "description": "No-regression baseline for src/services references to crate::server or super::server server-layer modules.",
   "captured_at": "2026-06-08",
-  "total_count": 24,
+  "total_count": 22,
   "files": {
     "src/services/auto_queue/phase_gate_violations.rs": {
       "count": 1
@@ -20,9 +20,6 @@
     "src/services/discord/recovery_engine.rs": {
       "count": 1
     },
-    "src/services/discord/runtime_bootstrap.rs": {
-      "count": 1
-    },
     "src/services/discord/turn_bridge/completion_guard.rs": {
       "count": 5
     },
@@ -30,7 +27,7 @@
       "count": 8
     },
     "src/services/dispatches/discord_delivery/orchestration.rs": {
-      "count": 2
+      "count": 1
     },
     "src/services/onboarding/channel.rs": {
       "count": 1
