+++ src/server/routes/dispatches/mod.rs
@@ -16,6 +16,8 @@ pub(crate) use outbox::dispatch_outbox_loop;
 pub use outbox::resolve_channel_alias_pub;
 pub(crate) use outbox::use_counter_model_channel;
 
-// ── Re-exports: Thread reuse ─────────────────────────────────
-pub(crate) use thread_reuse::validate_channel_thread_maps_on_startup_with_backends;
+// ── Re-exports: Thread reuse route handlers ──────────────────
+// The Postgres/Discord-API thread-map helpers were relocated to
+// `crate::services::dispatches::discord_delivery` (#3037); only the axum route
+// handlers remain in the route layer.
 pub use thread_reuse::{get_card_thread, get_pending_dispatch_for_thread, link_dispatch_thread};
