+++ src/services/dispatches/discord_delivery/mod.rs
@@ -1,5 +1,6 @@
 mod guard;
 mod orchestration;
+mod thread_reuse;
 mod transport;
 
 pub(crate) use guard::send_dispatch_with_delivery_guard;
@@ -10,6 +11,11 @@ pub(crate) use orchestration::{
     send_dispatch_to_discord_with_transport, send_review_result_to_primary,
     send_review_result_to_primary_with_transport, sync_dispatch_status_reaction,
     sync_dispatch_status_reaction_with_pg,
+};
+pub(crate) use thread_reuse::{
+    get_mapped_thread_for_channel_pg, get_thread_for_channel_pg,
+    set_thread_for_channel_map_only_pg, set_thread_for_channel_pg,
+    validate_channel_thread_maps_on_startup_with_backends,
 };
 pub(crate) use transport::{
     DispatchMessagePostError, DispatchMessagePostErrorKind, DispatchMessagePostOutcome,
